Change mapped_addr to be a 32 bit integer to prevent overflow in MBCs

+++ include/mbc_1.h
@@ -8,8 +8,8 @@ public:
     ~MBC1() = default;
 
 public:
-    bool read(uint16_t addr, uint16_t &mapped_addr, bool &rom_read) override;
-    bool write(uint16_t addr, uint8_t data, uint16_t &mapped_addr) override;
+    bool read(uint16_t addr, uint32_t &mapped_addr, bool &rom_read) override;
+    bool write(uint16_t addr, uint8_t data, uint32_t &mapped_addr) override;
 
     uint8_t getRAMBank();
     uint8_t getROMBank();
